TAIR/requirements.txt
xformers==0.0.25.post1
omegaconf==2.3.0
accelerate==0.28.0
einops==0.7.0
opencv_python==4.9.0.80
scipy==1.12.0
ftfy==6.2.0
regex==2023.12.25
python-dateutil==2.9.0.post0
timm==0.9.16
pytorch-lightning==2.2.1 # only for loading pretrained sd weight
tensorboard==2.16.2 # for tensorboard event visualization
protobuf==4.25.3 # for tensorboard
lpips==0.1.4
facexlib==0.3.0
gradio==4.43.0
polars==1.12.0
torchsde==0.2.6
bitsandbytes==0.44.1
wandb==0.19.8
pyiqa==0.1.13

# requirements for llava
transformers==4.37.2
tokenizers==0.15.1
sentencepiece==0.1.99

# requirements for ram
fairscale==0.4.4
